fix(colorpicker): resolver always returns null
During usage the graphql responses for the colorpicker field would always return null despite a value existing in the database. This was due to the processOutput hook using the old format from before an API change, this has since been resolved and the graphql queries now return their corresponding values.