feat(core,admin): add document full text search
Add a full text search on the document table to allow for simple yet powerful searching of documents.

This is achieved by creating a ViewEntity with the additional `fts` column, this may need to be revised into a trigger based strategy later on depending on scalability.

Additionally adds required UI components and log for utilising the full text search which is displayed on both documents and schema document listing pages.