feat: overhaul monorepo and improve separation of concerns
Overhaul the monorepo by separating database and graphql entities into their own package so they may be consumed by other packages rather than providing a separate type definition which is prone to being out of date with current.

In addition reduce the arguments the DockiteFields by passing the entire typeorm instance for use across fields which benefit from the added extraction of entities.

Finally extract field management and schema management to a package so they can be managed separately to the rest of the core codebase.