fix(admin): manage apollo cache hits
Currently the apollo cache is proving to be a burden with a number of cache hits occurring during critical updates or deletions requiring a full page reload to evict them. To manage this we are now following the best practices for the apollo cache and calling `resetStore` during any destructive activities until we can use `evict` with `@apollo/client:3.0.0`.