feat: token refreshing
A user will no longer be logged out after an hour if they were active. Instead shortlived (15m) tokens are issued along with a HttpOnly refresh token that is valid for 3 days. During usage if a users token expires a refresh will be attempted using the token in the cookie and then reissued via the authorization header.

Additionally removes `@nuxtjs/apollo` due to the inability to apply the required logic to implement refreshing on the frontend.