feat: add user management system
Add the user management system responsible for performing CRUD actions
on users and their corresponding roles. Handles the addition and removal
of scopes and resetting of passwords for users.

During a user creation or password reset action an email will be sent to
the users provided email address with their details.