feat(admin,core): add api keys
Add API keys to dockite allowing for a user to use the graphql api
without being required to follow the typical authentication flow with
handling for refreshing etc.

An api key can be created from the account page provided the user has
the correct scopes which can then be used against the api by providing
`x-dockite-token` as either a request header or query param.