feat: add graphql schema to dockite fields for usage
Add the current GraphQL schema in usage across the application
to the constructor arguments for a given dockite field.

This will provide reliable access to the schema for performing
internal queries for accessing data that has been processed by
dockite.

Notable usage includes the accessing of complex fields that are only
recognised during the output step of retrieving a document making it
impossible to use with the raw data stored within the database.